修复等式两边符号类型不一致的告警问题

+++ frameworks/eventhandler/src/event_handler.cpp
@@ -37,7 +37,7 @@ namespace {
 static constexpr int FFRT_SUCCESS = 0;
 static constexpr int FFRT_ERROR = -1;
 static constexpr int FFRT_TASK_REMOVE_FAIL = 1;
-static constexpr int64_t MILLISECONDS_TO_NANOSECONDS_RATIO = 1000000;
+static constexpr uint64_t MILLISECONDS_TO_NANOSECONDS_RATIO = 1000000;
 static const uint64_t PENDING_JOB_TIMEOUT[3] = {
     system::GetIntParameter("const.sys.notification.pending_higher_event_vip", 4),
     system::GetIntParameter("const.sys.notification.pending_higher_event_immediate", 40),
@@ -537,7 +537,7 @@ bool EventHandler::HasPendingHigherEvent(int32_t priority)
             return true;
         }
         if (eventHandleTime + PENDING_JOB_TIMEOUT[i] * MILLISECONDS_TO_NANOSECONDS_RATIO <=
-            now.time_since_epoch().count()) {
+            static_cast<uint64_t>(now.time_since_epoch().count())) {
             return true;
         }
     }
